fix: Add serializer to UserLoginView
